Split the API documentation tests by subject
One file had grown past the size we keep test files under. The package
endpoints, the admin CRUD endpoints, the rules every endpoint follows,
and the example listing itself now each have their own file, with the
shared readers beside them.
